refactor(search-service): update features/search/service imports
