+++ public/app/features/search/service/dummy.ts
@@ -1,7 +1,7 @@
 import { SelectableValue, DataFrame, DataFrameView } from '@grafana/data';
 import { TermCount } from 'app/core/components/TagFilter/TagFilter';
 
-import { GrafanaSearcher, QueryResponse, SearchQuery } from '.';
+import { GrafanaSearcher, QueryResponse, SearchQuery } from './types';
 
 // This is a dummy search useful for tests
 export class DummySearcher implements GrafanaSearcher {
