+++ public/app/features/search/service/sql.ts
@@ -7,10 +7,8 @@ import { PermissionLevelString } from 'app/types';
 import { DEFAULT_MAX_VALUES, GENERAL_FOLDER_UID, TYPE_KIND_MAP } from '../constants';
 import { DashboardSearchHit, DashboardSearchItemType } from '../types';
 
-import { LocationInfo } from './types';
+import { DashboardQueryResult, GrafanaSearcher, LocationInfo, QueryResponse, SearchQuery } from './types';
 import { replaceCurrentFolderQuery } from './utils';
-
-import { DashboardQueryResult, GrafanaSearcher, QueryResponse, SearchQuery } from '.';
 
 interface APIQuery {
   query?: string;
